+++ .eslintrc.js
@@ -1,23 +1,19 @@
 module.exports = {
-    "env": {
-        "es2021": true,
-        "node": true
-    },
-    "extends": [
-        "eslint:recommended",
-        "plugin:@typescript-eslint/recommended"
-    ],
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "ecmaVersion": "latest",
-        "sourceType": "module"
-    },
-    "plugins": [
-        "@typescript-eslint"
-    ],
-    "rules": {
-        "@typescript-eslint/ban-types": 'off',
-        '@typescript-eslint/no-explicit-any': 'off'
-    },
-    "ignorePatterns": ["example/*", "tests/**/*"]
+	env: {
+		es2021: true,
+		node: true
+	},
+	extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
+	parser: '@typescript-eslint/parser',
+	parserOptions: {
+		ecmaVersion: 'latest',
+		sourceType: 'module'
+	},
+	plugins: ['@typescript-eslint'],
+	rules: {
+		'@typescript-eslint/ban-types': 'off',
+		'@typescript-eslint/no-explicit-any': 'off',
+		'no-mixed-spaces-and-tabs': 'off'
+	},
+	ignorePatterns: ['example/*', 'tests/**/*']
 }